+++ PedidoApp/src/main/java/com/example/PedidoApp/controller/PedidoController/PedidoController.java
@@ -43,7 +43,7 @@ public class PedidoController {
 
 
     @GetMapping("/traerporId/{id}")
-    public ResponseEntity<Object> traerProducto(@PathVariable Long id) {
+    public ResponseEntity<Object> traerProducto(@PathVariable Integer id) {
 
         return new ResponseHandler().generateResponse(HttpStatus.OK, pedidoService.traerPorId(id));
 
@@ -51,7 +51,7 @@ public class PedidoController {
 
 
     @GetMapping("/generate-pdf/{id}")
-    public void generarPdfVenta(HttpServletResponse response, @PathVariable Long id) throws Exception {
+    public void generarPdfVenta(HttpServletResponse response, @PathVariable Integer id) throws Exception {
         System.out.println("sadfdddddddddddddddddddd");
         ventaReports.ventaReports(response, pedidoService.traerPorId(id));
