+++ PedidoApp/src/main/java/com/example/PedidoApp/repository/BodegaRepository/BodegaRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface BodegaRepository extends JpaRepository<Bodega, Long> {
+public interface BodegaRepository extends JpaRepository<Bodega, Integer> {
 }
