+++ PedidoApp/src/main/java/com/example/PedidoApp/repository/Pedido/PedidoRepository.java
@@ -3,7 +3,7 @@ package com.example.PedidoApp.repository.Pedido;
 import com.example.PedidoApp.model.Pedido;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PedidoRepository extends JpaRepository<Pedido, Long> {
+public interface PedidoRepository extends JpaRepository<Pedido, Integer> {
